requirements.txt
flask
flask-cors
langchain
langchain-core
langchain-community
langchain-ollama
langchain-groq
langchain-huggingface
langchain-text-splitters
sentence-transformers
faiss-cpu
requests
python-dotenv
pytest
